Fix tests, delete files related to experiments
